Add optional library for sqrt
We added an if-else clause to use our library rather than the stdlib for
sqrt. We first created an option in this file for USE_MYMATH variable.
This is set on as default. We can turn it off when we compile. We also
create other variables such as EXTRA_LIBS and EXTRA_INCLUDES that
include the appropriate flags for the linker to link in our own math
library and also headers.