Add #cmakedefine to config file
We defined a USE_MYMATH variable in our config file so that we can pass
those variables down to our source code. There was a question asking
what happens if we configure our config file before adding the option to
USE_MYMATH. If we invert the two (putting configure_file first) then we
create that header file and miss setting USE_MYMATH since it will not be
passed down to the source code.